Deep_EIoU/reid/requirements.txt
numpy==1.23.0
loguru
lap
Cython
h5py
Pillow
six
scipy
opencv-python
matplotlib
tb-nightly
future
yacs
gdown
flake8
yapf
isort==4.3.21
imageio
chardet
tabulate
pycocotools
thop
scikit-learn
numpy
seaborn
ultralytics
moviepy==1.0.3
ffmpeg
parse